chore(refactor): remove facade pattern
