+++ src/use-cases/changeName/index.js
@@ -1,47 +1,58 @@
 // eslint-disable-next-line no-unused-vars
-const ExampleRepository = require('../../../repositories/example.repository')
+const ExampleRepository = require('../../ports/secondary/example.repository')
 // eslint-disable-next-line no-unused-vars
-const Example = require('../../../entities/example')
+const Example = require('../../entities/example')
 // eslint-disable-next-line no-unused-vars
 const winston = require('winston')
+const IChangeName = require('../../ports/primary/changename')
 
-const ExampleNotFound = require('../../../exceptions/example-not-found')
+const ExampleNotFound = require('../../exceptions/example-not-found')
 
-class ChangeName {
+class ChangeName extends IChangeName {
   /**
    *
    * @param {ExampleRepository} repository
    * @param {winston.Logger} logger
    */
   constructor (repository, logger) {
+    super()
     this.repository = repository
     this.logger = logger
 
+    this.getLoggerContainer = this.getLoggerContainer.bind(this)
     this.execute = this.execute.bind(this)
   }
 
   /**
-   *
-   * @param {Example} example
+   * @param {string} id
+   * @param {string} name
    * @return {Promise<Example>}
    */
-  async execute (id, example) {
+  async execute (id, name) {
     this.logger.info('Changing the example name')
     // REPOSITORY
     // Retrieve the entity with all data
-    const updatedExample = await this.repository.getById(id)
+    const example = await this.repository.getById(id)
 
-    if (!updatedExample) throw new ExampleNotFound()
+    // BUSINESS EXCEPTIONS
+    if (!example) throw new ExampleNotFound()
     // ENTITY LOGIC
     // Change only the necessary field in the useCase
-    updatedExample.changeName(example.name)
+    example.changeName(name)
     // REPOSITORY
     // Update the entity
-    await this.repository.update(updatedExample)
+    await this.repository.update(example)
 
     this.logger.info('The name has been successfully changed')
 
-    return updatedExample
+    return example
+  }
+
+  /**
+   * @return {winston.Logger}
+   */
+  getLoggerContainer () {
+    return this.logger
   }
 }
